adding mentalState links
