+++ conscious/conscious/MentalStateView.swift
@@ -20,7 +20,6 @@ import UIKit
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        print("OOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOO")
         initSubviews()
     }
