+++ conscious/conscious/Sector.swift
@@ -1,0 +1,16 @@
+//
+//  Sector.swift
+//  conscious
+//
+//  Created by William Johnson on 3/13/16.
+//  Copyright © 2016 Conscious World. All rights reserved.
+//
+
+import Foundation
+
+class Sector: NSObject {
+    var minValue: CGFloat?
+    var maxValue: CGFloat?
+    var midValue: CGFloat?
+    var sector: Int?
+}
